Use long segment indices for DASH
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183389701

+++ library/dash/src/androidTest/java/com/google/android/exoplayer2/source/dash/manifest/DashManifestParserTest.java
@@ -64,7 +64,7 @@ public class DashManifestParserTest extends InstrumentationTestCase {
         if (representation instanceof Representation.MultiSegmentRepresentation) {
           Representation.MultiSegmentRepresentation multiSegmentRepresentation =
               (Representation.MultiSegmentRepresentation) representation;
-          int firstSegmentIndex = multiSegmentRepresentation.getFirstSegmentNum();
+          long firstSegmentIndex = multiSegmentRepresentation.getFirstSegmentNum();
           RangedUri uri = multiSegmentRepresentation.getSegmentUrl(firstSegmentIndex);
           assertThat(
                   uri.resolveUriString(representation.baseUrl)
